feature: Add server idle timeout.
When server is idle for period of time specified in decorator, then it
will shut down automatically. The granularity is 10ms to start, but we
can change this later if needed. If CPU is a concern we may change out
the current socketserver-based implementation for one based on asyncio.

Fixes #1.